Release tooling: GoReleaser, keyless Sigstore signing, SBOMs, ghcr images, changelog
- .goreleaser.yaml: static binaries for linux/darwin amd64/arm64 and
  linux/armv7, reproducible (-trimpath, commit timestamp; verified
  byte-identical across two snapshot builds), tar archives with the
  manual, sha256 checksums, SPDX SBOMs, keyless cosign signature on the
  checksums, multi-arch ghcr.io/edward-b-1/opens3 image built from the
  same binaries and signed by digest, generated changelog and notes.
- .github/workflows/release.yml on v* tags, pinned actions by SHA,
  make ci as the gate.
- make release-check / release-snapshot (goreleaser via go run, no
  install); CHANGELOG.md for 0.1.0; docs/RELEASING.md; README and
  manual install and verification instructions.
- opens3 version prints Go version, platform and commit; the version is
  now wired into the server so the console and admin info report it.
- object: fix an int overflow on 32-bit that broke the armv7 build.

+++ internal/object/copy.go
@@ -95,7 +95,7 @@ func (s *Service) UploadPartCopy(ctx context.Context, in UploadPartCopyInput) (*
 		size = src.Range.End - src.Range.Start + 1
 	}
 	if size > MaxPartSize {
-		return nil, nil, s3err.New(s3err.InvalidRequest).WithMessage("The specified copy source is larger than the maximum allowable size for a copy source: %d", MaxPartSize)
+		return nil, nil, s3err.New(s3err.InvalidRequest).WithMessage("The specified copy source is larger than the maximum allowable size for a copy source: %d", int64(MaxPartSize))
 	}
 	p, err := s.UploadPart(ctx, UploadPartInput{Bucket: in.Bucket, Key: in.Key, UploadID: in.UploadID, PartNumber: in.PartNumber, Body: src.Body, Size: size, SSE: in.SSE})
 	if err != nil {
